Updated kyle's and added a new one called lifecycle
Added the video to Kyle's form. To use the kinect controls it needs to
be a wpf project so I created a new one. It is by no means done but the
tracking and video work on it. Still need to add the timer and all that
jazz but the basic skeleton is there. We will need to make another
window or possibly another app for the clinitian. I will add more to
this branch before saturday.